+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-## 0.0.7
-
-* [ ] Support `file://` in `@import "..."` and `![](...)` and `[](...)`.    
+* [ ] Add `latexEngine` config.